@@ -25,8 +25,32 @@ final class EnglishInflector implements InflectorInterface
         // Fourth entry: Whether the suffix may succeed a consonant
         // Fifth entry: singular suffix, normal
 
-        // bacteria (bacterium), criteria (criterion), phenomena (phenomenon)
-        ['a', 1, true, true, ['on', 'um']],
+        // bacteria (bacterium)
+        ['airetcab', 8, true, true, 'bacterium'],
+
+        // corpora (corpus)
+        ['aroproc', 7, true, true, 'corpus'],
+
+        // criteria (criterion)
+        ['airetirc', 8, true, true, 'criterion'],
+
+        // curricula (curriculum)
+        ['alucirruc', 9, true, true, 'curriculum'],
+
+        // genera (genus)
+        ['areneg', 6, true, true, 'genus'],
+
+        // media (medium)
+        ['aidem', 5, true, true, 'medium'],
+
+        // memoranda (memorandum)
+        ['adnaromem', 9, true, true, 'memorandum'],
+
+        // phenomena (phenomenon)
+        ['anemonehp', 9, true, true, 'phenomenon'],
+
+        // strata (stratum)
+        ['atarts', 6, true, true, 'stratum'],
 
         // nebulae (nebula)
         ['ea', 2, true, true, 'a'],
@@ -97,6 +121,9 @@ final class EnglishInflector implements InflectorInterface
         // statuses (status)
         ['sesutats', 8, true, true, 'status'],
 
+        // article (articles), ancle (ancles)
+        ['sel', 3, true, true, 'le'],
+
         // analyses (analysis), ellipses (ellipsis), fungi (fungus),
         // neuroses (neurosis), theses (thesis), emphases (emphasis),
         // oases (oasis), crises (crisis), houses (house), bases (base),
@@ -141,7 +168,7 @@ final class EnglishInflector implements InflectorInterface
         // shoes (shoe)
         ['se', 2, true, true, ['', 'e']],
 
-         // status (status)
+        // status (status)
         ['sutats', 6, true, true, 'status'],
 
         // tags (tag)
@@ -238,7 +265,10 @@ final class EnglishInflector implements InflectorInterface
         // teeth (tooth)
         ['htoot', 5, true, true, 'teeth'],
 
-        // bacteria (bacterium), criteria (criterion), phenomena (phenomenon)
+        // albums (album)
+        ['mubla', 5, true, true, 'albums'],
+
+        // bacteria (bacterium), curricula (curriculum), media (medium), memoranda (memorandum), phenomena (phenomenon), strata (stratum)
         ['mu', 2, true, true, 'a'],
 
         // men (man), women (woman)
@@ -247,20 +277,11 @@ final class EnglishInflector implements InflectorInterface
         // people (person)
         ['nosrep', 6, true, true, ['persons', 'people']],
 
-        // bacteria (bacterium), criteria (criterion), phenomena (phenomenon)
-        ['noi', 3, true, true, 'ions'],
+        // criteria (criterion)
+        ['noiretirc', 9, true, true, 'criteria'],
 
-        // coupon (coupons)
-        ['nop', 3, true, true, 'pons'],
-
-        // seasons (season), treasons (treason), poisons (poison), lessons (lesson)
-        ['nos', 3, true, true, 'sons'],
-
-        // icons (icon)
-        ['noc', 3, true, true, 'cons'],
-
-        // bacteria (bacterium), criteria (criterion), phenomena (phenomenon)
-        ['no', 2, true, true, 'a'],
+        // phenomena (phenomenon)
+        ['nonemonehp', 10, true, true, 'phenomena'],
 
         // echoes (echo)
         ['ohce', 4, true, true, 'echoes'],
@@ -270,6 +291,9 @@ final class EnglishInflector implements InflectorInterface
 
         // atlases (atlas)
         ['salta', 5, true, true, 'atlases'],
+
+        // aliases (alias)
+        ['saila', 5, true, true, 'aliases'],
 
         // irises (iris)
         ['siri', 4, true, true, 'irises'],
@@ -396,11 +420,11 @@ final class EnglishInflector implements InflectorInterface
 
         // aircraft
         'tfarcria',
+
+        // hardware
+        'erawdrah',
     ];
 
-    /**
-     * {@inheritdoc}
-     */
     public function singularize(string $plural): array
     {
         $pluralRev = strrev($plural);
@@ -432,7 +456,7 @@ final class EnglishInflector implements InflectorInterface
                 if ($j === $suffixLength) {
                     // Is there any character preceding the suffix in the plural string?
                     if ($j < $pluralLength) {
-                        $nextIsVowel = false !== strpos('aeiou', $lowerPluralRev[$j]);
+                        $nextIsVowel = str_contains('aeiou', $lowerPluralRev[$j]);
 
                         if (!$map[2] && $nextIsVowel) {
                             // suffix may not succeed a vowel but next char is one
@@ -477,9 +501,6 @@ final class EnglishInflector implements InflectorInterface
         return [$plural];
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function pluralize(string $singular): array
     {
         $singularRev = strrev($singular);
@@ -512,7 +533,7 @@ final class EnglishInflector implements InflectorInterface
                 if ($j === $suffixLength) {
                     // Is there any character preceding the suffix in the plural string?
                     if ($j < $singularLength) {
-                        $nextIsVowel = false !== strpos('aeiou', $lowerSingularRev[$j]);
+                        $nextIsVowel = str_contains('aeiou', $lowerSingularRev[$j]);
 
                         if (!$map[2] && $nextIsVowel) {
                             // suffix may not succeed a vowel but next char is one
