@@ -3,8 +3,8 @@
 /*
  * This file has been auto-generated by the Symfony String Component for internal use.
  *
- * Unicode version: 15.1.0
- * Date: 2023-09-13T11:47:12+00:00
+ * Unicode version: 16.0.0
+ * Date: 2024-09-11T08:21:22+00:00
  */
 
 return [
@@ -45,12 +45,20 @@ return [
         9749,
     ],
     [
+        9776,
+        9783,
+    ],
+    [
         9800,
         9811,
     ],
     [
         9855,
         9855,
+    ],
+    [
+        9866,
+        9871,
     ],
     [
         9875,
@@ -394,7 +402,7 @@ return [
     ],
     [
         12736,
-        12771,
+        12773,
     ],
     [
         12783,
@@ -451,6 +459,10 @@ return [
     [
         13312,
         19903,
+    ],
+    [
+        19904,
+        19967,
     ],
     [
         19968,
@@ -837,6 +849,10 @@ return [
         101589,
     ],
     [
+        101631,
+        101631,
+    ],
+    [
         101632,
         101640,
     ],
@@ -879,6 +895,14 @@ return [
     [
         110960,
         111355,
+    ],
+    [
+        119552,
+        119638,
+    ],
+    [
+        119648,
+        119670,
     ],
     [
         126980,
@@ -1054,23 +1078,19 @@ return [
     ],
     [
         129664,
-        129672,
+        129673,
     ],
     [
-        129680,
-        129725,
-    ],
-    [
-        129727,
-        129733,
+        129679,
+        129734,
     ],
     [
         129742,
-        129755,
+        129756,
     ],
     [
-        129760,
-        129768,
+        129759,
+        129769,
     ],
     [
         129776,
