@@ -3,8 +3,8 @@
 /*
  * This file has been auto-generated by the Symfony String Component for internal use.
  *
- * Unicode version: 15.1.0
- * Date: 2023-09-13T11:47:13+00:00
+ * Unicode version: 16.0.0
+ * Date: 2024-09-11T08:21:22+00:00
  */
 
 return [
@@ -109,7 +109,7 @@ return [
         2139,
     ],
     [
-        2200,
+        2199,
         2207,
     ],
     [
@@ -917,11 +917,15 @@ return [
         68903,
     ],
     [
+        68969,
+        68973,
+    ],
+    [
         69291,
         69292,
     ],
     [
-        69373,
+        69372,
         69375,
     ],
     [
@@ -1045,6 +1049,26 @@ return [
         70516,
     ],
     [
+        70587,
+        70592,
+    ],
+    [
+        70606,
+        70606,
+    ],
+    [
+        70608,
+        70608,
+    ],
+    [
+        70610,
+        70610,
+    ],
+    [
+        70625,
+        70626,
+    ],
+    [
         70712,
         70719,
     ],
@@ -1122,6 +1146,10 @@ return [
     ],
     [
         71453,
+        71453,
+    ],
+    [
+        71455,
         71455,
     ],
     [
@@ -1277,12 +1305,24 @@ return [
         73538,
     ],
     [
+        73562,
+        73562,
+    ],
+    [
         78912,
         78912,
     ],
     [
         78919,
         78933,
+    ],
+    [
+        90398,
+        90409,
+    ],
+    [
+        90413,
+        90415,
     ],
     [
         92912,
@@ -1399,6 +1439,10 @@ return [
     [
         124140,
         124143,
+    ],
+    [
+        124398,
+        124399,
     ],
     [
         125136,
