@@ -111,16 +111,20 @@ class XliffFileLoader implements LoaderInterface
                     continue;
                 }
 
-                if (isset($translation->target) && 'needs-translation' === (string) $translation->target->attributes()['state']) {
+                $source = (string) (isset($attributes['resname']) && $attributes['resname'] ? $attributes['resname'] : $translation->source);
+
+                if (isset($translation->target)
+                    && 'needs-translation' === (string) $translation->target->attributes()['state']
+                    && \in_array((string) $translation->target, [$source, (string) $translation->source], true)
+                ) {
                     continue;
                 }
 
-                $source = isset($attributes['resname']) && $attributes['resname'] ? $attributes['resname'] : $translation->source;
                 // If the xlf file has another encoding specified, try to convert it because
                 // simple_xml will always return utf-8 encoded values
                 $target = $this->utf8ToCharset((string) ($translation->target ?? $translation->source), $encoding);
 
-                $catalogue->set((string) $source, $target, $domain);
+                $catalogue->set($source, $target, $domain);
 
                 $metadata = [
                     'source' => (string) $translation->source,
@@ -143,7 +147,7 @@ class XliffFileLoader implements LoaderInterface
                     $metadata['id'] = (string) $attributes['id'];
                 }
 
-                $catalogue->setMetadata((string) $source, $metadata, $domain);
+                $catalogue->setMetadata($source, $metadata, $domain);
             }
         }
     }
