@@ -166,11 +166,11 @@ function extractLocaleFromFilePath($filePath)
 function extractTranslationKeys($filePath)
 {
     $translationKeys = [];
-    $contents = new \SimpleXMLElement(file_get_contents($filePath));
+    $contents = new SimpleXMLElement(file_get_contents($filePath));
 
     foreach ($contents->file->body->{'trans-unit'} as $translationKey) {
         $translationId = (string) $translationKey['id'];
-        $translationKey = (string) $translationKey->source;
+        $translationKey = (string) ($translationKey['resname'] ?? $translationKey->source);
 
         $translationKeys[$translationId] = $translationKey;
     }
